Move object_to_json to GcObject
